fix(training-activities-template): change distance column type from int to float

+++ apps/runmates-back/src/app/training-plan-templates/entities/training-activities-template.entity.ts
@@ -12,8 +12,8 @@ export class TrainingActivityTemplateEntity {
  @Column()
  activity: string;
 
- @Column({ type: 'int', nullable: true })
- distance: number;
+  @Column({ type: 'float', nullable: true })
+  distance: number;
 
  @Column({ nullable: true })
  description: string;
